refactor(serializers): Move base serializers to core app
Relocate base serializers for users, promos, and comments from the `user` and `business` apps to the central `core` app. This change promotes code reuse and establishes a more logical and scalable project structure.

- Moved `BaseUserSerializer`, `BaseUserPromoSerializer`, and `BaseCommentSerializer` to `core/serializers.py`.
- Renamed `BasePromoSerializer` to `BaseCompanyPromoSerializer` for clarity.
- Updated all dependent serializers in the `user` and `business` apps to inherit from the new centralized base classes in `core`.
