+++ promo_code/user/serializers.py
@@ -1,4 +1,3 @@
-import django.contrib.auth.password_validation
 import django.core.cache
 import django.db.transaction
 import rest_framework.exceptions
@@ -8,67 +7,12 @@ import rest_framework_simplejwt.token_blacklist.models as tb_models
 import rest_framework_simplejwt.tokens
 
 import business.constants
-import business.models
 import core.serializers
 import core.utils.auth
-import user.constants
 import user.models
 
 
-class OtherFieldSerializer(rest_framework.serializers.Serializer):
-    age = rest_framework.serializers.IntegerField(
-        required=True,
-        min_value=user.constants.AGE_MIN,
-        max_value=user.constants.AGE_MAX,
-    )
-    country = core.serializers.CountryField(required=True)
-
-
-class BaseUserSerializer(rest_framework.serializers.ModelSerializer):
-    password = rest_framework.serializers.CharField(
-        write_only=True,
-        required=True,
-        validators=[django.contrib.auth.password_validation.validate_password],
-        max_length=user.constants.PASSWORD_MAX_LENGTH,
-        min_length=user.constants.PASSWORD_MIN_LENGTH,
-        style={'input_type': 'password'},
-    )
-    name = rest_framework.serializers.CharField(
-        required=True,
-        min_length=user.constants.NAME_MIN_LENGTH,
-        max_length=user.constants.NAME_MAX_LENGTH,
-    )
-    surname = rest_framework.serializers.CharField(
-        required=True,
-        min_length=user.constants.SURNAME_MIN_LENGTH,
-        max_length=user.constants.SURNAME_MAX_LENGTH,
-    )
-    email = rest_framework.serializers.EmailField(
-        required=True,
-        min_length=user.constants.EMAIL_MIN_LENGTH,
-        max_length=user.constants.EMAIL_MAX_LENGTH,
-    )
-    avatar_url = rest_framework.serializers.URLField(
-        required=False,
-        max_length=user.constants.AVATAR_URL_MAX_LENGTH,
-        allow_null=True,
-    )
-    other = OtherFieldSerializer(required=True)
-
-    class Meta:
-        model = user.models.User
-        fields = (
-            'name',
-            'surname',
-            'email',
-            'password',
-            'avatar_url',
-            'other',
-        )
-
-
-class SignUpSerializer(BaseUserSerializer):
-
+class SignUpSerializer(core.serializers.BaseUserSerializer):
     @django.db.transaction.atomic
     def create(self, validated_data):
         try:
@@ -153,7 +97,7 @@ class SignInSerializer(
         return token
 
 
-class UserProfileSerializer(BaseUserSerializer):
+class UserProfileSerializer(core.serializers.BaseUserSerializer):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         for field in self.fields.values():
@@ -251,77 +195,7 @@ class UserFeedQuerySerializer(
         return attrs
 
 
-class BaseUserPromoSerializer(rest_framework.serializers.ModelSerializer):
-    """
-    Base serializer for promos, containing common fields and methods.
-    """
-
-    promo_id = rest_framework.serializers.UUIDField(source='id')
-    company_id = rest_framework.serializers.UUIDField(source='company.id')
-    company_name = rest_framework.serializers.CharField(source='company.name')
-    active = rest_framework.serializers.BooleanField(source='is_active')
-    like_count = rest_framework.serializers.IntegerField(
-        source='get_like_count',
-    )
-    comment_count = rest_framework.serializers.IntegerField(
-        source='get_comment_count',
-    )
-
-    is_liked_by_user = rest_framework.serializers.SerializerMethodField()
-    is_activated_by_user = rest_framework.serializers.SerializerMethodField()
-
-    class Meta:
-        model = business.models.Promo
-        fields = (
-            'promo_id',
-            'company_id',
-            'company_name',
-            'description',
-            'image_url',
-            'active',
-            'is_activated_by_user',
-            'like_count',
-            'comment_count',
-            'is_liked_by_user',
-        )
-        read_only_fields = fields
-
-    def get_is_liked_by_user(self, obj: business.models.Promo) -> bool:
-        """
-        Checks whether the current user has liked this promo.
-        """
-        request = self.context.get('request')
-        if (
-            request
-            and hasattr(request, 'user')
-            and request.user.is_authenticated
-        ):
-            return user.models.PromoLike.objects.filter(
-                promo=obj,
-                user=request.user,
-            ).exists()
-        return False
-
-    def get_is_activated_by_user(self, obj: business.models.Promo) -> bool:
-        """
-        Checks whether the current user has activated this promo code.
-        """
-        request = self.context.get('request')
-
-        if not (
-            request
-            and hasattr(request, 'user')
-            and request.user.is_authenticated
-        ):
-            return False
-
-        return user.models.PromoActivationHistory.objects.filter(
-            promo=obj,
-            user=request.user,
-        ).exists()
-
-
-class PromoFeedSerializer(BaseUserPromoSerializer):
+class PromoFeedSerializer(core.serializers.BaseUserPromoSerializer):
     """
     Serializer for representing promo feed data for a user.
     """
@@ -329,7 +203,7 @@ class PromoFeedSerializer(BaseUserPromoSerializer):
     pass
 
 
-class UserPromoDetailSerializer(BaseUserPromoSerializer):
+class UserPromoDetailSerializer(core.serializers.BaseUserPromoSerializer):
     """
     Serializer for detailed promo-code information
     (without revealing the code value).
@@ -339,33 +213,18 @@ class UserPromoDetailSerializer(BaseUserPromoSerializer):
     pass
 
 
-class UserAuthorSerializer(BaseUserSerializer):
+class UserAuthorSerializer(core.serializers.BaseUserSerializer):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         for field in self.fields.values():
             field.required = False
 
-    class Meta(BaseUserSerializer.Meta):
+    class Meta(core.serializers.BaseUserSerializer.Meta):
         fields = ('name', 'surname', 'avatar_url')
         read_only_fields = fields
 
 
-class BaseCommentSerializer(rest_framework.serializers.ModelSerializer):
-    """
-    Base serializer for promo comments.
-    """
-
-    text = rest_framework.serializers.CharField(
-        min_length=user.constants.COMMENT_TEXT_MIN_LENGTH,
-        max_length=user.constants.COMMENT_TEXT_MAX_LENGTH,
-    )
-
-    class Meta:
-        model = user.models.PromoComment
-        fields = ('text',)
-
-
-class CommentSerializer(BaseCommentSerializer):
+class CommentSerializer(core.serializers.BaseCommentSerializer):
     """
     Serializer for displaying (reading) a comment.
     """
@@ -378,19 +237,19 @@ class CommentSerializer(BaseCommentSerializer):
     )
     author = UserAuthorSerializer(read_only=True)
 
-    class Meta(BaseCommentSerializer.Meta):
-        fields = BaseCommentSerializer.Meta.fields + (
+    class Meta(core.serializers.BaseCommentSerializer.Meta):
+        fields = core.serializers.BaseCommentSerializer.Meta.fields + (
             'id',
             'date',
             'author',
         )
 
 
-class CommentCreateSerializer(BaseCommentSerializer):
+class CommentCreateSerializer(core.serializers.BaseCommentSerializer):
     pass
 
 
-class CommentUpdateSerializer(BaseCommentSerializer):
+class CommentUpdateSerializer(core.serializers.BaseCommentSerializer):
     pass
